requirements.txt
# ─── LangChain Core ────────────────────────────────────────────────────────────
langchain>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.2.0
langchain-text-splitters>=0.3.0

# ─── Vector Store ──────────────────────────────────────────────────────────────
chromadb>=0.5.0

# ─── Embeddings ────────────────────────────────────────────────────────────────
sentence-transformers>=3.0.0

# ─── Document Ingestion ────────────────────────────────────────────────────────
pypdf>=4.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# ─── Token Counting ────────────────────────────────────────────────────────────
tiktoken>=0.7.0

# ─── Configuration ─────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
pyyaml>=6.0.0

# ─── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-mock>=3.14.0
